# HttpExample/18.HttpHealthcheck/obj/Debug/net8.0/EndpointInfo/7.Swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Todo API",
    "description": "간단한 Todo 관리용 Web API",
    "contact": {
      "name": "김ss",
      "email": "you@example.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/Todo/{id}": {
      "get": {
        "tags": [
          "Todo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Todo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Todo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GetTodoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "UpdateTodoDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "isComplete": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}